help

+++ strp.c
@@ -77,7 +77,7 @@ char* Str_search(const char *haystack, const char *needle)
   if(Str_getLength(needle) == 0) return (char*)(haystack);
 
   /*check determines if we have found a potential needle*/
-  int check = 0;
+  size_t check = 0;
   char *last_occurence;
   const char *startNeedle = needle;
